requirements.txt
absl-py==2.1.0
appdirs==1.4.4
attrs==23.2.0
av==12.0.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressai==1.2.4
contourpy==1.1.1
coolname==2.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
facenet-pytorch==2.5.3
flatbuffers==24.3.25
fonttools==4.51.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.16.4
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.8
install==1.3.5
iopath==0.1.10
joblib==1.4.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.5
lightning-utilities==0.11.2
lpips==0.1.4
matplotlib==3.6.3
mediapipe==0.10.3
moviepy==1.0.3
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.8.0.76
opencv-python==4.7.0.68
packaging==24.0
pandas==1.5.2
pathtools==0.1.2
Pillow==9.4.0
piq==0.8.0
plotille==5.0.0
plotly==5.17.0
portalocker==2.7.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==3.20.3
psnr-hvsm==0.1.0
psutil==5.9.8
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-msssim==0.2.1
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
runx==0.0.11
scikit-image==0.19.3
scikit-learn==1.2.0
scikit-video==1.1.11
scipy==1.10.1
seaborn==0.13.0
sentry-sdk==1.45.0
setproctitle==1.3.3
siti-tools==0.2.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.4.12
torch==1.13.1
torch-dct==0.1.6
torchfile==0.1.0
torchmetrics==1.2.1
torchstat==0.0.7
torchvision==0.14.1
tqdm==4.65.2
typing_extensions==4.11.0
urllib3==2.2.1
visdom==0.1.8.9
Vizer==0.1.5
wandb==0.13.9
wcwidth==0.2.13
websocket-client==1.4.0
yacs==0.1.8
